Add verbs to results and make code clearer
Gone are the days of staring at some bitfields and wondering
just what the fuck they do.  It was clever but obscure, and
relied on careful bean counting in order for it to work.
Array.prototype.includes() is a much nicer solution, even
considering that I have to type strings repeatedly; that is
hardly an issue compared to seeing what the fucking code does.